refactor(leagues-frontend): Move LocationSelect to standalone component
